% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_games.R
\name{annual_games}
\alias{annual_games}
\title{Annual games}
\usage{
annual_games(data, year_value)
}
\arguments{
\item{data}{a data set}

\item{year_value}{a year of interest}
}
\value{
total number of games played for the given year
}
\description{
This function returns the total number of games played for a given year for the NBA player Giannis Antentokounmpo.
}
\examples{
annual_games()
}
