% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_numeric.R
\name{cor_numeric}
\alias{cor_numeric}
\title{Cor numeric}
\usage{
cor_numeric(data, year_value)
}
\arguments{
\item{data}{a data set}

\item{year_value}{a year of interest}
}
\value{
a correlation plot of all the numeric variables of a given data set
}
\description{
This function selects all numeric variables in a data set and return the correlation plot of the variables
}
\examples{
cor_numeric()
}
