% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_choice.R
\name{graph_choice}
\alias{graph_choice}
\title{Creates a plot of points, for, against, and difference over time for a team}
\usage{
graph_choice(choice, team_name, df)
}
\arguments{
\item{team_name}{The name of a team \code{inputParameter1}}

\item{df}{data fram containg college basketball games played \code{inputParameter2}}
}
\value{
Graph of points over time
}
\description{
Takes a team name and tibble of college basketball games played and returns
either a plot of their points for and against, or a plot of their points diffences
over time depending on the choice argument. The choice argument must be "for_against"
or "difference". This function was created to make the shiny app work the way it was
desired to.
}
\examples{
graph_choice("difference","Gonzaga", cbb)
}
\keyword{keywords}
