% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenpom_reformat.R
\name{kenpom_reformat}
\alias{kenpom_reformat}
\title{Reformats the text file of college basketball games played from kenpom}
\usage{
kenpom_reformat(cbb)
}
\arguments{
\item{cbb}{A tibble of college basketball games played}
}
\value{
A more useful version on the games played for a shiny app
}
\description{
Takes the text file of games played, and returns a more pretty and complete
version of it. It counts points difference in the games, and was created
for use of a shiny app.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}
