% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nba}
\alias{nba}
\title{Data that contains the aggregate statistics of 67 NBA seasons}
\format{
A data set with 24,691 rows and 53 variables
}
\source{
https://www.kaggle.com/drgilermo/nba-players-stats/data?select=Seasons_Stats.csv
}
\usage{
nba
}
\description{
Contains 53 player statistics from the NBA seasons
}
\examples{
data(nba)
}
\keyword{datasets}
