% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{Oldest player}
\usage{
oldest_player(year_value)
}
\arguments{
\item{year_value}{a year of interest}
}
\value{
oldest player for a given year
}
\description{
This function  returns the oldest player for any given year, if there is a tie, then alphabetical order is used as a deciding factor.
}
\examples{
oldest_player()
}
