% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_bar.R
\name{points_bar}
\alias{points_bar}
\title{Creates a plot of points, for, against, and difference over time for a team}
\usage{
points_bar(team_name, df)
}
\arguments{
\item{team_name}{The name of a team \code{inputParameter1}}

\item{df}{data fram containg college basketball games played \code{inputParameter2}}
}
\value{
Graph of points over time
}
\description{
Takes a team name and tibble of college basketball games played and returns
a bar graph of their points per game and points allowed per game.
}
\examples{
points_graph("Gonzaga", cbb)
}
\keyword{keywords}
