% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_diff_hist.R
\name{points_diff_hist}
\alias{points_diff_hist}
\title{Creates a plot of points, for, against, and difference over time for a team}
\usage{
points_diff_hist(team_name, df)
}
\arguments{
\item{team_name}{The name of a team \code{inputParameter1}}

\item{df}{data fram containg college basketball games played \code{inputParameter2}}
}
\value{
Graph of points over time
}
\description{
Takes a team name and tibble of college basketball games played and returns
a histogram of their points differnces in their games.
}
\examples{
points_graph("Gonzaga", cbb)
}
\keyword{keywords}
