% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_graph.R
\name{points_graph}
\alias{points_graph}
\title{Creates a plot of points, for, against, and difference over time for a team}
\usage{
points_graph(team_name, df)
}
\arguments{
\item{team_name}{The name of a team \code{inputParameter1}}

\item{df}{data fram containg college basketball games played \code{inputParameter2}}
}
\value{
Graph of points over time
}
\description{
Takes a team name and tibble of college basketball games played and returns
a plot with dates on the y axis and points on the x axis with 3 seperate lines.
A green line representing points for, a red line representing poins against, and
a blue line representing points difference.
}
\examples{
points_graph("Gonzaga", cbb)
}
\keyword{keywords}
