% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Runs the shiny app}
\usage{
run_app()
}
\value{
runs the shiny app for the basketball package
}
\description{
Runs the shiny app in the basketball package
}
\examples{
/dontrun{run_app()}
}
