% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_games.R
\name{team_games}
\alias{team_games}
\title{Filters college basketball data to one specific team}
\usage{
team_games(team_name, df)
}
\arguments{
\item{team_name}{The name of a college basketball team \code{inputParameter1}}

\item{df}{The data frame that contains all the college basketball games played \code{inputParameter2}}
}
\value{
Returns only the games played by the specified team
}
\description{
Using data of college basketball games played from kenpom.com
this function takes the name of a team and a tibble of all games played
and returns games played by that team.Change score_diff from home-away to
positive if the given team won, and negative if it lost.
}
\examples{
Southern_Utah games <- team_games("Southern Utah", cbb)
}
\keyword{keywords}
