% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_graph.R
\name{win_graph}
\alias{win_graph}
\title{Creates a plot of points, for, against, and difference over time for a team}
\usage{
win_graph(team_name, df)
}
\arguments{
\item{team_name}{The name of a team \code{inputParameter1}}

\item{df}{data fram containg college basketball games played \code{inputParameter2}}
}
\value{
Graph of points over time
}
\description{
Takes a team name and tibble of college basketball games played and returns
a bar graph of wins and losses.
A green line representing points for, a red line representing poins against, and
a blue line representing points difference.Requires ggplot2 package
}
\examples{
points_graph("Gonzaga", cbb)
}
\keyword{keywords}
